change data type for ParSortID

+++ src/TestProblem/Hydro/Plummer/FB_Plummer.cpp
@@ -15,7 +15,7 @@ extern double Plummer_FB_Like;
 
 // function pointers to be set by FB_Init_Plummer()
 extern int (*FB_User_Ptr)( const int lv, const double TimeNew, const double TimeOld, const double dt,
-                           const int NPar, const int *ParSortID, real_par *ParAtt[PAR_NATT_TOTAL],
+                           const int NPar, const long *ParSortID, real_par *ParAtt[PAR_NATT_TOTAL],
                            real (*Fluid)[FB_NXT][FB_NXT][FB_NXT], const double EdgeL[], const double dh, bool CoarseFine[],
                            const int TID, RandomNumber_t *RNG );
 extern void (*FB_End_User_Ptr)();
@@ -87,7 +87,7 @@ extern void (*FB_End_User_Ptr)();
 // Return      :  Fluid, ParAtt
 //-------------------------------------------------------------------------------------------------------
 int FB_Plummer( const int lv, const double TimeNew, const double TimeOld, const double dt,
-                const int NPar, const int *ParSortID, real_par *ParAtt[PAR_NATT_TOTAL],
+                const int NPar, const long *ParSortID, real_par *ParAtt[PAR_NATT_TOTAL],
                 real (*Fluid)[FB_NXT][FB_NXT][FB_NXT], const double EdgeL[], const double dh, bool CoarseFine[],
                 const int TID, RandomNumber_t *RNG )
 {
@@ -123,7 +123,7 @@ int FB_Plummer( const int lv, const double TimeNew, const double TimeOld, const 
 
    for (int t=0; t<NPar; t++)
    {
-      const int    p      = ParSortID[t];
+      const long   p      = ParSortID[t];
       const double xyz[3] = { (double)ParAtt[PAR_POSX][p], (double)ParAtt[PAR_POSY][p], (double)ParAtt[PAR_POSZ][p] };
 
       int idx[3];
